diplomacy: AddressMapEntry and BindingScope.collect
